+++ pyaerocom/aeroval/_processing_base.py
@@ -3,8 +3,10 @@ import abc
 from pyaerocom._lowlevel_helpers import TypeValidator
 from pyaerocom.aeroval import EvalSetup
 from pyaerocom.aeroval.experiment_output import ExperimentOutput
+from pyaerocom.aeroval.json_utils import round_floats
 from pyaerocom.colocation.colocation_setup import ColocationSetup
 from pyaerocom.colocation.colocator import Colocator
+from pyaerocom.utils import recursive_defaultdict, recursive_defaultdict_to_dict
 
 
 class HasConfig:
@@ -69,6 +71,119 @@ class ProcessingEngine(HasConfig, abc.ABC):
 
         """
         pass
+
+    def _add_heatmap_timeseries_entry(
+        self,
+        entry: dict,
+        region: str,
+        network: str,
+        obsvar: str,
+        layer: str,
+        modelname: str,
+        modvar: str,
+    ):
+        """Adds a heatmap entry to hm/ts
+
+        :param entry: The entry to be added.
+        :param network: Observation network
+        :param obsvar: Observation variable
+        :param layer: Vertical layer
+        :param modelname: Model name
+        :param modvar: Model variable
+        """
+        project = self.exp_output.proj_id
+        experiment = self.exp_output.exp_id
+
+        with self.avdb.lock():
+            glob_stats = self.avdb.get_heatmap_timeseries(
+                project, experiment, region, network, obsvar, layer, default={}
+            )
+            glob_stats = recursive_defaultdict(glob_stats)
+            glob_stats[obsvar][network][layer][modelname][modvar] = round_floats(entry)
+            self.avdb.put_heatmap_timeseries(
+                recursive_defaultdict_to_dict(glob_stats),
+                project,
+                experiment,
+                region,
+                network,
+                obsvar,
+                layer,
+            )
+
+    def _add_heatmap_entry(
+        self,
+        entry,
+        frequency: str,
+        network: str,
+        obsvar: str,
+        layer: str,
+        modelname: str,
+        modvar: str,
+    ):
+        """Adds a heatmap entry to glob_stats
+
+        :param entry: The entry to be added.
+        :param region: The region (eg. ALL)
+        :param obsvar: Observation variable.
+        :param layer: Vertical Layer (eg. SURFACE)
+        :param modelname: Model name
+        :param modelvar: Model variable.
+        """
+        project = self.exp_output.proj_id
+        experiment = self.exp_output.exp_id
+
+        with self.avdb.lock():
+            glob_stats = self.avdb.get_glob_stats(project, experiment, frequency, default={})
+            glob_stats = recursive_defaultdict(glob_stats)
+            glob_stats[obsvar][network][layer][modelname][modvar] = round_floats(entry)
+            self.avdb.put_glob_stats(
+                recursive_defaultdict_to_dict(glob_stats), project, experiment, frequency
+            )
+
+    def _write_station_data(self, data):
+        """Writes timeseries weekly.
+
+        :param data: Data to be written.
+        """
+        project = self.exp_output.proj_id
+        experiment = self.exp_output.exp_id
+
+        location = data["station_name"]
+        network = data["obs_name"]
+        obsvar = data["var_name_web"]
+        layer = data["vert_code"]
+        modelname = data["model_name"]
+        with self.avdb.lock():
+            station_data = self.avdb.get_timeseries_weekly(
+                project, experiment, location, network, obsvar, layer, default={}
+            )
+            station_data[modelname] = round_floats(data)
+            self.avdb.put_timeseries_weekly(
+                station_data, project, experiment, location, network, obsvar, layer
+            )
+
+    def _write_timeseries(self, data):
+        """Write timeseries"""
+        if not isinstance(data, list):
+            data = [data]
+
+        project = self.exp_output.proj_id
+        experiment = self.exp_output.exp_id
+        with self.avdb.lock():
+            for d in data:
+                location = d["station_name"]
+                network = d["obs_name"]
+                obsvar = d["var_name_web"]
+                layer = d["vert_code"]
+                modelname = d["model_name"]
+
+                timeseries = self.avdb.get_timeseries(
+                    project, experiment, location, network, obsvar, layer, default={}
+                )
+                timeseries[modelname] = round_floats(d)
+                self.avdb.put_timeseries(
+                    timeseries, project, experiment, location, network, obsvar, layer
+                )
 
 
 class HasColocator(HasConfig):
