+++ tests/aeroval/test_coldatatojson_helpers.py
@@ -6,18 +6,13 @@ from pathlib import Path
 import numpy as np
 from pytest import mark, param
 
-from pyaerocom.aeroval.coldatatojson_helpers import (
-    _add_heatmap_entry_json,
+from pyaerocom.aeroval.coldatatojson_helpers import (  # _add_heatmap_entry_json,; _write_diurnal_week_stationdata_json,; _write_site_data,; _write_stationdata_json,; get_stationfile_name,
     _init_stats_dummy,
     _prepare_aerocom_regions_json,
     _prepare_country_regions,
     _prepare_default_regions_json,
     _prepare_htap_regions_json,
     _prepare_regions_json_helper,
-    _write_diurnal_week_stationdata_json,
-    _write_site_data,
-    _write_stationdata_json,
-    get_stationfile_name,
 )
 from pyaerocom.region import get_all_default_region_ids
 from pyaerocom.region_defs import (
@@ -27,24 +22,23 @@ from pyaerocom.region_defs import (
     OTHER_REGIONS,
 )
 
-
-def test__write_stationdata_json(tmp_path: Path):
-    data = dict(station_name="stat1", obs_name="obs1", var_name_web="var1", vert_code="Column")
-    path: Path = tmp_path / get_stationfile_name(**data)
-    assert not path.exists()
-
-    data["model_name"] = "model1"
-    _write_stationdata_json(data, path.parent)
-    assert path.exists()
-
-    data["model_name"] = "model2"
-    _write_stationdata_json(data, path.parent)
-    assert path.exists()
-
-    data = json.loads(path.read_text())
-    assert data["model1"].pop("model_name") == "model1"
-    assert data["model2"].pop("model_name") == "model2"
-    assert data["model1"] == data["model2"]
+# def test__write_stationdata_json(tmp_path: Path):
+#    data = dict(station_name="stat1", obs_name="obs1", var_name_web="var1", vert_code="Column")
+#    path: Path = tmp_path / get_stationfile_name(**data)
+#    assert not path.exists()
+#
+#    data["model_name"] = "model1"
+#    _write_stationdata_json(data, path.parent)
+#    assert path.exists()
+#
+#    data["model_name"] = "model2"
+#    _write_stationdata_json(data, path.parent)
+#    assert path.exists()
+#
+#    data = json.loads(path.read_text())
+#    assert data["model1"].pop("model_name") == "model1"
+#    assert data["model2"].pop("model_name") == "model2"
+#    assert data["model1"] == data["model2"]
 
 
 # def test__write_site_data(tmp_path: Path):
@@ -61,45 +55,43 @@ def test__write_stationdata_json(tmp_path: Path):
 #    assert not list(tmp_path.glob("*.json"))
 #    _write_site_data(data, str(tmp_path))
 #    assert len(list(tmp_path.glob("*.json"))) == len(data)
+
+# def test__write_diurnal_week_stationdata_json(tmp_path: Path):
+#    data = dict(station_name="stat1", obs_name="obs1", var_name_web="var1", vert_code="Column")
+#    dirs = {"ts/diurnal": tmp_path}
+#    path: Path = tmp_path / get_stationfile_name(**data)
+#    assert not path.exists()
 #
+#    data["model_name"] = "model1"
+#    _write_diurnal_week_stationdata_json(data, dirs)
+#    assert path.exists()
 #
-def test__write_diurnal_week_stationdata_json(tmp_path: Path):
-    data = dict(station_name="stat1", obs_name="obs1", var_name_web="var1", vert_code="Column")
-    dirs = {"ts/diurnal": tmp_path}
-    path: Path = tmp_path / get_stationfile_name(**data)
-    assert not path.exists()
+#    data["model_name"] = "model2"
+#    _write_diurnal_week_stationdata_json(data, dirs)
+#    assert path.exists()
+#
+#    data = json.loads(path.read_text())
+#    assert data["model1"].pop("model_name") == "model1"
+#    assert data["model2"].pop("model_name") == "model2"
+#    assert data["model1"] == data["model2"]
 
-    data["model_name"] = "model1"
-    _write_diurnal_week_stationdata_json(data, dirs)
-    assert path.exists()
-
-    data["model_name"] = "model2"
-    _write_diurnal_week_stationdata_json(data, dirs)
-    assert path.exists()
-
-    data = json.loads(path.read_text())
-    assert data["model1"].pop("model_name") == "model1"
-    assert data["model2"].pop("model_name") == "model2"
-    assert data["model1"] == data["model2"]
-
-
-def test__add_heatmap_entry_json(tmp_path: Path):
-    path = tmp_path / "add_entry.json"
-    assert not path.exists()
-
-    result = dict(value=None)
-    obs, obs_var, vert, model, model_var = "obs", "obs_var", "Column", "model", "model_var"
-    _add_heatmap_entry_json(path, result, obs, obs_var, vert, model, model_var)
-    assert path.exists()
-
-    data = json.loads(path.read_text())
-    assert data[obs_var][obs][vert][model][model_var] == result
-
-    result.update(value=42.0)
-    _add_heatmap_entry_json(path, result, obs, obs_var, vert, model, model_var)
-
-    data = json.loads(path.read_text())
-    assert data[obs_var][obs][vert][model][model_var] == result
+# def test__add_heatmap_entry_json(tmp_path: Path):
+#    path = tmp_path / "add_entry.json"
+#    assert not path.exists()
+#
+#    result = dict(value=None)
+#    obs, obs_var, vert, model, model_var = "obs", "obs_var", "Column", "model", "model_var"
+#    _add_heatmap_entry_json(path, result, obs, obs_var, vert, model, model_var)
+#    assert path.exists()
+#
+#    data = json.loads(path.read_text())
+#    assert data[obs_var][obs][vert][model][model_var] == result
+#
+#    result.update(value=42.0)
+#    _add_heatmap_entry_json(path, result, obs, obs_var, vert, model, model_var)
+#
+#    data = json.loads(path.read_text())
+#    assert data[obs_var][obs][vert][model][model_var] == result
 
 
 def test__init_stats_dummy():
