+++ tests/aeroval/test_coldatatojson_helpers2.py
@@ -6,7 +6,7 @@ import pytest
 import xarray
 
 from pyaerocom import ColocatedData, TsType
-from pyaerocom.aeroval.coldatatojson_helpers import (
+from pyaerocom.aeroval.coldatatojson_helpers import (  # get_heatmap_filename,; get_json_mapname,; get_stationfile_name,; get_timeseries_file_name,
     _create_diurnal_weekly_data_object,
     _get_jsdate,
     _get_period_keys,
@@ -16,33 +16,28 @@ from pyaerocom.aeroval.coldatatojson_helpers import (
     _map_indices,
     _process_statistics_timeseries,
     _remove_less_covered,
-    get_heatmap_filename,
-    get_json_mapname,
     get_profile_filename,
-    get_stationfile_name,
-    get_timeseries_file_name,
 )
 from pyaerocom.aeroval.exceptions import TrendsError
 from pyaerocom.exceptions import TemporalResolutionError, UnknownRegion
 from tests.fixtures.collocated_data import COLDATA
 
-
-def test_get_heatmap_filename():
-    assert get_heatmap_filename("daily") == "glob_stats_daily.json"
-
-
-def test_get_timeseries_filename():
-    assert get_timeseries_file_name("reg1", "obs1", "var1", "vert1") == "reg1-obs1-var1-vert1.json"
-
-
-def test_get_stationfile_name():
-    json = get_stationfile_name("stat1", "obs1", "var1", "Column")
-    assert json == "stat1_obs1-var1_Column.json"
-
-
-def test_get_json_mapname():
-    json = get_json_mapname("obs1", "var1", "mod1", "var1", "Column", "period")
-    assert json == "obs1-var1_Column_mod1-var1_period.json"
+# def test_get_heatmap_filename():
+#    assert get_heatmap_filename("daily") == "glob_stats_daily.json"
+#
+#
+# def test_get_timeseries_filename():
+#    assert get_timeseries_file_name("reg1", "obs1", "var1", "vert1") == "reg1-obs1-var1-vert1.json"
+#
+#
+# def test_get_stationfile_name():
+#    json = get_stationfile_name("stat1", "obs1", "var1", "Column")
+#    assert json == "stat1_obs1-var1_Column.json"
+#
+#
+# def test_get_json_mapname():
+#    json = get_json_mapname("obs1", "var1", "mod1", "var1", "Column", "period")
+#    assert json == "obs1-var1_Column_mod1-var1_period.json"
 
 
 def get_profile_filename():
